move autocomplete-input.tsx
